+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+- Kill language server zombie processes
+
 ## 1.16.3
 
 - Bump Swift-MesonLSP to v3.1.1